+++ modules/aws-http4s/src/smithy4s/aws/internals/AwsJsonCodecs.scala
@@ -53,6 +53,7 @@ private[aws] object AwsJsonCodecs {
       .withErrorBodyDecoders(jsonDecoders)
       .withErrorDiscriminator(AwsErrorTypeDecoder.fromResponse(jsonDecoders))
       .withRequestMediaType(contentType)
+      .withSmithyPathEncoding(false)
   }
 
 }