+++ modules/json/test/src/smithy4s/json/JsonSpec.scala
@@ -21,6 +21,7 @@ import smithy4s.schema.Schema
 import smithy4s.Blob
 import smithy4s.Document
 import Schema._
+import smithy4s.example.DefaultTest
 
 class JsonSpec() extends FunSuite {
 
@@ -86,6 +87,11 @@ class JsonSpec() extends FunSuite {
 
     assertEquals(result, expectedJson)
     assertEquals(roundTripped, Right(foo))
+  }
+
+  test("Json read - optional field with a default value accepts nulls") {
+    val result = Json.read[DefaultTest](Blob("""{"two": null}"""))
+    assertEquals(result, Right(DefaultTest()))
   }
 
 }